RNP-126: Refactor errorcodes to int's instead of String

+++ android/src/main/java/com/onegini/mobile/sdk/reactnative/handlers/customregistration/SimpleCustomRegistrationActionImpl.kt
@@ -30,7 +30,7 @@ class SimpleCustomRegistrationActionImpl(private val idProvider: String) : Onegi
             customRegistrationCallback.returnError(exception)
             callback = null
         } ?: throw OneginiReactNativeException(
-            OneginiWrapperErrors.REGISTRATION_NOT_IN_PROGRESS.code.toInt(),
+            OneginiWrapperErrors.REGISTRATION_NOT_IN_PROGRESS.code,
             OneginiWrapperErrors.REGISTRATION_NOT_IN_PROGRESS.message
         )
     }
