+++ android/src/main/java/com/onegini/mobile/sdk/reactnative/handlers/pins/CreatePinRequestHandler.kt
@@ -35,7 +35,7 @@ class CreatePinRequestHandler : OneginiCreatePinRequestHandler {
         pinCallback?.let { callBack ->
             callBack.acceptAuthenticationRequest(pin)
             return true
-        } ?: throw OneginiReactNativeException(OneginiWrapperErrors.REGISTRATION_NOT_IN_PROGRESS.code.toInt(), OneginiWrapperErrors.REGISTRATION_NOT_IN_PROGRESS.message)
+        } ?: throw OneginiReactNativeException(OneginiWrapperErrors.REGISTRATION_NOT_IN_PROGRESS.code, OneginiWrapperErrors.REGISTRATION_NOT_IN_PROGRESS.message)
     }
 
     @Throws(OneginiReactNativeException::class)
@@ -43,6 +43,6 @@ class CreatePinRequestHandler : OneginiCreatePinRequestHandler {
         pinCallback?.let { callback ->
             callback.denyAuthenticationRequest()
             pinCallback = null
-        } ?: throw OneginiReactNativeException(OneginiWrapperErrors.REGISTRATION_NOT_IN_PROGRESS.code.toInt(), OneginiWrapperErrors.REGISTRATION_NOT_IN_PROGRESS.message)
+        } ?: throw OneginiReactNativeException(OneginiWrapperErrors.REGISTRATION_NOT_IN_PROGRESS.code, OneginiWrapperErrors.REGISTRATION_NOT_IN_PROGRESS.message)
     }
 }
